requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
requests >= 2.21
mongoengine==0.16
django-allauth >= 0.39.1
django-mongoengine-forms >= 0.4.6
django-cors-headers
djangorestframework >= 3.9.4
django-rest-framework-mongoengine >= 3.3.1
gunicorn >= 19.9.0
